+++ src/omniscript/__init__.py
@@ -3,4 +3,4 @@
 from .api import OmniEngine, RunResult
 
 __all__ = ["OmniEngine", "RunResult", "__version__"]
-__version__ = "0.2.0"
+__version__ = "0.2.1"